Fix non-'open' channels not being monitored
Fix #85
raidenInitializationEpic was supposed to fire initial batch of events,
including `channelMonitored` events (to monitor events for all current
channels), but there was a filter which caused it to fire this event
only for channels currently in 'open' state. This prevented other events
(like 'ChannelSettled') to be picked later if the app was reloaded.
This commit extends raidenInitializationEpic tests to ensure
'channelMonitored' is emitted even for 'settling' channels (currently
last channel state), ensuring any channel event will be properly picked.