Add chainId and registry address to state
This will be useful when trying to validate/ensure a loaded JSON file
corresponds to a given TokenNetwork. Currently, this checking is done
implicitly by having these info in the localStorage key, but this only
works if one uses the localStorage overload of the `create` factory
(which isn't the only way to load a state, and in the future we'll
support downloading and uploading a new/backup state).