Add smart parameter validation for all routers
Parameter provided by the user must be validated. The goal was to keep
the router description easy to read, while having the full power of
validation. Therefore it uses describing names for validation chains,
while keeping them configurable at the same time. The validation itself
is imperative to simplify the middleware stack for the routes as well as
make it better readable.