Extend controller logic for channel routes
This includes mainly the creation of new channels and retrieving open
channels with different filters applied. All routes attach their
channels as result to the request. A shared middleware convert their
format according to the API specification and send them to the user. The
format transformation might not be 100% correct at the moment due to
data model differences. They must be analyzed and resolved.

Furthermore has the shutdown process of the Raiden service been
improved. Thereby it is able (again) to stop on receiving error events.