cli: fix shutdown procedure for raiden stop
When the user cancels the running application, the shutdown procedure
gets triggered. It stops the REST server and also stops the raiden
instance. Depending on still open connections etc, this might take
a while. If the user sends a second cancel signal, the whole process
should be killed without waiting anymore. This was the former behavior
that has been restored now.