cli: rework implementation structure to be flat
The following design decision has been made: The raiden sdk should be
not wrapped and provided by a singleton. Rather it should be a global
variable that becomes attached to functions by their context. Therefore
each module explicitly exports a "make" function. This allows us to bind
or call them with the global context, including all instances they need.
Thereby the raiden instance (also the logger as another example) gets
passed down up to the routes of the server. The key advantage is that it
is now possible to modify the global context at any place, but take
effect everywhere. Moreover does this functional approach allow to test
the components easily on their own. It is so modular, that it can be
easily re-used for any other use-case rather than start it from the
command-line.

As addition to that, it has been decided that the routes should include
the main logic of the endpoints. If someone likes to checkout a simple
reference implementation for the sdk for a specific case, he can simply
checkout the according route. Therefore I just exported utility
functions that are used across the routes, making code reusable and
easier to read.

The parameter validation of the re-implemented channel endpoints have
been removed for now. It is affected by another major design decision,
part of upcoming commits. Related to that, it is not possible to create
new channels for this commit.