cli: remove/reduce user parameter validation
It has been decided that the cli should validate as less user parameters
as possible. The goal is to let the sdk doing the heavy lifting,
including the parameter validation. The API layer then simply parses
possibly thrown errors, including wrong parameter (formats). This will
become mixed with the other errors for cases where the balances is not
enough and similar cases.
The remaining validation is for cases where the sdk can not do the work
and also should not. An example would be the filtering of channels by
token and partner address. The API specification requires to send error
codes for male formatted address parameter. Empty response lists are
actual valid results.