cli: refactor/start full compatible Raiden API implementation
The focus of the cli has changed to provide a full compatible implementation of the REST API specification by Raiden. Therefore the so far provided "shortcut" functions via the cli arguments have been removed, focusing on running the server only.
The current implementation includes the scaffold for the whole application, including all routes the specification defines. Only a few of the endpoints are implemented yet, more will follow soon.
On the way a few major design decisions have been made and fixed in a scoped CONTRIBUTING guide. Due to that, the implementations has become refactored multi times on its way to the current state, to provide the best fundament we can.