sdk: improve parameter validation on open channel
According to the Raiden API specification, the errors must be
identifiable to become grouped and mapped to a specific error code.
Therefore the decode function has been adjusted to allow the provision
of a custom error. The scheme follows the strategy of the custom assert
function in a more simplified manner.

Moreover have some new error codes been added to improve the specificity
of the SDKs errors.