cli: reduce code complexity on channels router
This was triggered by a report of codeclimate. Due to the design
decision that the routes should export per default a function
explicitly so that is can become bound to a context, the channel router
(as currently only implemented router) become to big. So the number of
lines that act as a complexity parameter was too big.
Therefore the router middleware functions have become exported to
standalone functions instead of arrow functions. They again get bound to
the global context. An additionally gained advantage is that the router
definitions reads more nice now.