dapp: build for different modes on CircleCI
There is a need to have the dapp build in different modes and provide
them as artifacts by CircleCI. At the moment only the build for GH pages
is available.
This commit makes the dapp build available as custom command that gets
parametrized. Thereby it can be re-used by multiple jobs. The default
setup is configured for the `ghpages` mode. The default workflow is now
also building an artifact in development mode. This is meant to be used
for PR reviews.

Moreover has the compression format changed. Since the `.tar.xz` is
rarely used. The `tar.gz` version is used more widely and therefore the
preferred choice.