dapp: move busy indicator in address input field
The address input field component has an indicator that show its busy
state. This is for example the case when it tries to resolve an ENS
name. This indicator was placed at the front of the fields text. This
causes the text to jump left and back again for the time the indicator
appears.
To work into the favor of the reworked spinner component the indicator
has been moved to the end/appendix of the input. There it replaces the
QR code icon while input field is busy. Thereby the text does not jump
anymore and the spinner components remains more straight forward.

Related to: #1773