sdk: refactor pathFindServiceEpic for clean code
The reactive epics of the SDK are huge. They are not rarely longer than
200 lines of code in a single function. Their indentation level goes up
to 10, resulting into many split statements. As consequence the epics
lack in having abstractions layers to help the reader understanding the
higher level idea. They are implemented everything from the top to the
bottom with all details. A new reader gets easily lost, wasting hours in
trying to understand what is going on.
Therefore I decided to cleanup the pathFindServiceEpic which is target
of a new feature. The code is 100% functional, still using a declarative
style of behavior. Furthermore I tried to introduce a few of the
heuristics for clean code:
  - functions names should say what they do
  - functions should descend only one level of abstraction
  - choose descriptive names including side-effects
  - inappropriate/redundant informative comments
  - understand the algorithm
  - obvious behavior is unimplemented
  - replace magic numbers with named constants