dapp: prevent unintended/automatic unloading/stop
There were issues when the web-browser automatically unloads the dApp.
Especially Firefox seem to do this for memory optimization if the tab
was unused for a longer period. This leads to a stop of the Raiden
service and the user needs to reconnect. This gets prevented now.

NOTE: This feature/fix is incredibly hard to test, because it is now
reliable reproducible and costs a takes long. But long term test
attempts (over 40h) seem to show that the unload preventing
functionality works.

Related to: #1805