sdk: extend pathFindServiceEpic to secure presence
There is the issue that the SDK considers the target of a route to be
offline if the presence is unknown. The returned error message is
therefore actually wrong.This required the workaround by the user to
manually do first a presence request of the target before doing
a transfer. This is not intuitive and confusing for future users.
Therefore the SDK has been extended to make sure that the targets
presence is actually known before evaluating a route. Therefore it
checks if the presence is already known, else it is doing a matrix
presence requests and waits for its response first before proceed.

Related to: #1835