sdk: accept falsie optional paymentId parameter
The optional paymentId parameter for a transfer call was checked for
presence by being truly. Else an identifier gets generated
automatically. This lead to the inconsistent behavior, that for example
a string `"0"` gets accepted and parsed as number zero, but proving `0`
as number directly got ignored. Therefore the origin absence of
a parameter is tested now, which is undefined. If a user provides an
invalid input, the decoding of it will throw an error as usual.

Related to: #1882