dapp: Simplify how AppHeader retrieves title
Every route provides a title by its meta-data. The same goes already for
the HomeRoute. But for reasons the HomeRoute was the only which got
a title local for i18n and had a special case in the AppHeader to use
this translation instead. the meta-data and the local string where the
same.
Since the conditional title was determined by the connection state, this
will cause issues for the upcoming disclaimer route, which takes place
before the home route. Therefore this got cleaned up and simplified to
simply use the meta data only.