dapp: Add first simple version of the disclaimer
Due to legal reasons there should be a disclaimer shown before the user
can reach the connect screen. Therefore a new route has been added to
show this disclaimer and let the user accept it. Therefore it becomes
the default route, navigating to the home route.
Note that the new route is not loaded in lazy manner as many other
routes. Since it is the new default route it should be part of the main
bundle. Furthermore should the home route remain not lazy as well, since
in future development the user can also land on the home screen if he
had accepted the disclaimer already once.