dapp: Prevent routing before disclaimer accepted
If the user has not yet accepted the disclaimer, he is not allowed to
route anywhere directly (e.g. via a known URL). Instead he gets
redirected to the disclaimer route.
Therefore the fact that the user accepted the disclaimer once must be
stored in the state of the app. This is then used as guard to all other
routes. The flag gets set when the user checks the checkbox and click
the accept button.