dapp: Simplify app header when on disclaimer route
There could be theoretically (rarely practically) the case that the
disclaimer gets shown when the app had already connected. This will
enable components in the header which should not be accessible. But
in general the header should be empty (except the title) if the
disclaimer is shown. Therefore the render conditions have been extended.

This includes a small optimization of the old code. It was a pattern
that the header components are rendered conditionally by the same
boolean expression. So it was an optimization as well making the code
better readable to export this into a computed value. It turned out that
this expression was the same as the store getter isConnected. So this
was simplified in the template, since the header was already mapping
this getter into the component.