dapp: Add feature to redirect after disclaimer
If the user tries to access a part of the app before accepting the
disclaimer, he becomes redirected to the disclaimer by a route guard.
Unfortunately this will forget about the original location the user
wanted to access (e.g. open channel with hub). This is now preserved as
a route query parameter. The disclaimer will then route to this location
if such is given. Else if navigates to the home route as default.

Note that we need to save the full path as navigation target to preserve
also the route parameters the user eventually has set (e.g. token
and partner address).