dapp: Prevent reset of disclaimer flag in store
The store gets reset when the app connects. This will loose the flag
that the user has accepted the disclaimer. Thereby the route guard to
protect app access without having accepted the disclaimer popped up
again. To fix this, the part of state that gets preserved during a reset
(so far only the settings) got extended to include this flag.