dapp: add option to persist disclaimer acceptance
The user should be able to say that his decision to accept the
disclaimer should be permanently saved. As a result he can access any
routes directly in a new session, if he has done so in any former
session. Furthermore does the landing page load directly into the home
route and skip the disclaimer route.