dapp: improve unit test with new store utils
It was a widely spread pattern to connect an account to the app for many
test cases. This was done by some commits to the store. These lines
where copy-pasted everywhere. This has two downsides. First these lines
of code do not express what their intention as test preparation is for.
And second does it bloat the test cases with repeating code. Test cases
should be short in preparation and focus on what they want to test.
Therefore the new store utils have been introduced, providing such
a preparation function for connecting an account. This has been adopted
everywhere.

Furthermore was there a misplaced test suite in the utils directory.
Therefore it simply became moved one layer up.