dapp: fix prop typing for transaction component
The component to display a transaction is expecting to get the
transaction object as property. This property was written to be of type
Transfers. Note already the final 's'. The Transfers type is defined to
be a dictionary of transaction keys to RaidenTransfer objects (be aware
that the terms transaction and transfer are used exchangeable). But the
template of the component was using this property as a plain
RaidenTransfer object.
Therefore the type of this property got fixed. Unfortunately this was
undetected. We need to check how we can improve the type checks with
Vue.