dapp: introduce test data generator utility
Test cases and suites should not care about generating test data. Rather
they should simply be able to use test data and adjust it to their needs
of the specific test case. Therefore a new test utility to generate such
data has been introduced. It is mainly focused on generating tokens and
transfer objects (including random Ethereum addresses). They are generic
enough to allow the user (test case) to partially overwrite these
objects during creation to match the specific test scenario.
As first "user" of this generator, the unit tests to display
a transaction have been adopted to this pattern. This is in preparation
for the upcoming transaction list component tests.

Furthermore should unit tests try to maintain dependencies to other
units. Such include the Vuex store for many Vue components. Rather the
capabilities of the vue-test-utils package should be used. These allow
the component wrapper to mock the store object and inject the data
necessary for the tests.