dapp: add unit test for transaction list component
There were no unit tests for this list so far. Since the functionality
has been recently extended, it was necessary to finally add those tests.
They include cases for the old/already existing functionality, as well
as the new filter feature.

Note that there is something awkward on components using the lazy
listing feature of Vuetify. For some reason does the 'mount' wrapper
function not render the list entries. But the 'shallowMount' wrapper
does. This is unexpected behavior and got therefore also documented into
the code.