sdk: update integration test image dependencies
Related to the update of the Raiden contracts to version v0.37.1, the
integration image was required to get updated as well. Therefore the
latest Raiden Python client that still support the chosen contracts is
used now. Furthermore the Raiden services have been updated as well to
be conform with the contracts.

Note that some comments in the Dockerfile have been removed. They are
duplicates of the related documentation. The variable names themselves
should be descriptive enough else.