Update and improve docs for test environment
The documentation how to upgrade the documentation was out of date. This
primarily was causes by a change to the hub.raiden.network resource
setup that does not use the Ansible Tower anymore.

Furthermore I attempted to cleanup the documentation as
a first-time-reader. I used example versions according to the current
ongoing upgrade to provide a better idea how the things look like in
real-world.