Update and improve docs for integration tests
The documentation that describe how to use the Docker the image for
integration tests were in need of an update. This was mainly caused
by some updates of the Raiden TransCore team. While following the steps
to update the image for a new version of the contracts and all related
dependencies I attempted to improve them with the collected knowledge
during the process.

A tiny improvement is to stick with a fixed order of the services
running in the image. This includes the port publishing when running the
Docker container.

Extend the descriptions of the services running for someone new to the
project.

Add documentation and elaborate the differences between the two
different Raiden contract version and how two find two compatible ones.
This was quite confusing, since the contracts themselves are still on
the old version, but the Python package has got updated.