dApp: UDC deposit bug fixes (#2088)
* Removes id

Removing the id prevents warning about two duplicate ids in the DOM when the amount input component is used on the main routes and the account screen routes simultaneously. The id is never used anyways.

* Add utility token to list of tokens

We need to add the utility token to the list of tokens in order to update its balance on each new block so that the most recent balance can be accessed from the state.

* Display current UDC token balance and format numbers

Get the UDC token balance from the store which is now updated on each new block and use a filter to format the balance.

* Changelog entry

* Removing assignment of displayFormat

Makes no sense to first assign displayFormat to a variable when it can be used directly from import.