cli: remove obsolete entrypoint
There have evolved two different entrypoints to start the CLI. It has
been decided to only keep the newer one and get rid of the old script.

Deleting this file become its own commit so the git history is more easy
to read. The intention is to rename the new entrypoint script called
'raiden' to the default name 'index'. Since this is the name of the old
entrypoint, this would end up in a huge diff, rather then a deleted and
a renamed file.