cli: rename entrypoint script to default again
According to the former commit the old default entrypoint was removed.
Only the newer one called 'raiden' was renaming. This has now become the
new default entry again.
Note that this commit was split up to have a cleaner git history.

The raiden executable was still handling the old case of having a built
bundle. This does not exist anymore and was removed. Therefore the
script become more simplified again, not having to treat this case.