dapp: remove buggy LazyRoute component
The LazyRoute component was meant to check if the dApp is connected or
not and display the Home route to let the user connect again if not. The
LazyRoute was wrapping all routes that require the user to be connected.
This Route lead to the bug that when reloading a route (the connection
gets lost during this process) the then displayed Home route was using
the title of the actual route wrapped/hidden in the LazyRoute. So the
Vue router was actual not aware that the Home route was shown.
Furthermore does this approach lead to a mass of duplicated code. So it
required to introduce the separation of view and navigation components.
The view components are the actual routes used by the router. Each of
them was basically a copy-paste to have the LazyRoute wrapping the
actual component. That was then a navigation component. The same goes
for the tests. All unit tests for the view components where copy-paste
to check if the wrapped component gets shown when being connected.
Even this is a single feature, it was required to get retest it for
each view.

Such feature like protecting routes when the user is not
connected/logged in/... is a common pattern that gets usually solved via
route guards. Such get globally defined at the Vue router. This then
provides a single place where the logic is defined and gets tested. The
views themselves can simply show their content (what is in the
navigation components) without caring about the connection.

Therefore the following changes have been done. All of them are
following the same pattern. There has been nothing new added, also the
VCS difference will look huge.
Each navigation component got removed and their content directly placed
into the according view component. The views original code got thereby
deleted. The exact same applied to the tests. The test cases for the
navigation components moved to the route test suites. Finally the
LazyRoute itself has become deleted.
Future commits will recover the guarding of the routes to require the
user to be connected.