dApp: Transfer screen redirect when no channels exist (#2103)
* Adding dialog for no open channels

A simple dialog to be displayed when user has no tokens with open channels. It has an warning icon and a button which redirects the user to the screen for selecting a new token.

* Replaces v-card-actions with a div

No real reason to use v-card-actions which happens to break the layout. Therefore replacing it with a simple div.

* Test for NoChannelsDialog component

* openChannels getter and trigger dialog in Transfer component if no open channels

Adding getter for returning a boolean if any open channels are available. This getter is used in the Transfer component to trigger the NoChannelsDialog if no channels are available. I don't think we need a watcher, mapGetters should be sufficient.

* Changelog entry