dApp: Channel settlement notifications improvement (#2112)
* Updated notification card design

The notification card now has a "proper" design. For the description, the text is split into an array of words which is gets iterated over in the template. Any word that happens to be an address gets wrapped in the address display component.

* Extracting notification description display to separate component

It will be reused for the snackbar.

* Shortened description

We want to limit the description to one line per notification and therefore need to exlcude some information.

* Notification panel animation

This adds a nice little sliding animation when displaying and hiding the notification panel.

* Only display notification title

This component has been significantly scaled down. There is no way a user could take in all information that used to be displayed within 5 seconds. All we need to display is the notifiction title to alert the user. Additional information can be found when accessing the notification panel.

* Make the container fit to the size of its content

* Various notification icons

* Icon type

* Icon as part of notification context enum

* Avoid splitting by word

* Padding to address only

Since we are no longer extracting word by word from the description the phrases that are extracted will already have whitespace between each word.

* RegEx more ethereum address specific

* Fixes failing NotificationCard tests

* Fixes failing tests for raiden-service

* Fixes failing tests for notifications area component

* Adjusting colors, locales and components related to notifications

* Changelog entry

* Removing icon from context and adding it as separate property