dapp: extend navigation guard for connection state
After the LazyRoute component has been removed, there was the need for
a new solution to make sure the user has connected the wallet. As
already pointed out this is a typical case for a navigation guard at the
router. Therefore the existing guard has been extended to not only
secure dApp access if the user has not accepted the disclaimer, but also
if the user has connected.
It was necessary to introduce navigation to the home view. This works
exactly as for the disclaimer view, acknowledging redirects via route
query parameter.