dapp: solve navigation issues on home view
There was the bug that if the user connects to the app while having
a query parameter to redirect after (e.g. using link to open channel
with hub), this was ignored and the user landed at the transfer view.
This was caused by two in parallel triggered navigations. First was the
correct redirect, executed by the home view itself. Second was caused by
the NoTokens component that redirects to the transfer view if the user
has any token. The NoTokens component was used within the home view,
after the user has connected.
Due to the new navigation guard for the connection state the old
functionality was not working anymore. In general is the design of
navigating via the NoTokens component pretty weird. So did it took
a while to finally find the source of navigation issues.
As solution the NoTokens component has been simplified to let the user
connect to a new token without any additional automated features. The
home view does not use it anymore and thereby got restricted to the
connection process only. So does also the redirection work properly
now. The NoTokens component has become integrated into the transfer
view, displayed when the user has no connected tokens.