dapp: refactor global navigation guard design
Due to the recent changes, the global navigation guard for the Vue
router got extended. While the guard itself was still good readable
enough, it was kinda impossible to adopt the unit tests for it. Already
the test case descriptions became incredibly long to cover the case its
state and result.
This is a good hint that the different parts of the global router
created a dependency that is hard to maintain. Changes are hard to
establish and test are incredibly hard to adjust, delete or extend.

Unfortunately it isn't to easy to split the navigation guard into
multiple parts. This is caused by the middleware architecture of the Vue
router. But because of the need, some kind of sub-middleware was
introduced. This design allows to split the global guard into multiple
smaller guards. Those are put into a list sorted by priority. Such child
guards can than be evaluated one by one. If one of them causes
a navigation, this gets executed. Else the global guard continues to the
next one. Finally if no guard is causing any navigation, the original
target gets navigated to.
This approach allows to have a clean modular code structure. This
structure does not only allow to have a more organized code base, but
also allows to write independent unit tests. So each guard function can
be tested individually and focused. Finally the global guards logic to
make use of a list of child guards needs to be tested.
After all it is now straight forward to remove guards, add new ones or
adopt a single one without affecting other parts of the routing.