dapp: fix infinite routing due to redirect query
If any component causes a navigation back to the disclaimer or home
route, this will cause an infinite loop of redirects. In this
state the disclaimer is accepted and the connection established, so both
child guards will cause a redirect to the "next" route. The issue is
that the global guard will then inject the original navigation target
as redirect parameter. This is the home or disclaimer route itself. And
the loop begins.
As the global guard could/should not know any details of the childrens,
the redirection functionality was shifted to the them.