dapp:  make transfer route more robust
The transfer route must can handle the case that the optional route
parameter for the token address is undefined. In this case it should
show the first known token the user has connected to.
This was requiring to rework/extend the unit tests for this component,
which were not properly setup before. This included a new data generator
function for channel objects.