dApp: blockchain progress for opening a channel (#2136)
* Expose channel open success event

This event will be needed for keeping track of when a channel gets opened.

* Preparations to pass target confirmation block number to channel open notification

Gets the values necessary for counting blocks from the channel open success event and passes them to a helper function for dispatching a notification whenever a channel open success event is registered.

* Refactoring notification store to either add or update a notification

One mutation for adding or updating a notification based on whether a notification is dispatched with an id that already exists in the notifications array.

* Update data passed to the notification success helper function

Notification success now also takes in the values which were added when refactoring the store to enable updating of notifications.

* Fixes bugs where id could reach infinity and wrongly typed notifications

* Moving all add and replace notifications logic to mutations

Calling the getter for incrementing the id inside of the actions caused the id to be incremented even when new notifications were dispatched with the same id. To solve this we had to refactor and move everything to the mutations.

* Adjusting Vue components in accordance to store refactoring

The NotificationCard component now displays the block count if the txConfirmationBlock is defined. Also, the NotificationArea component has been renamed to NotificationSnackbar since we are always refering to it as a snackbar.

* Fixes broken tests

Apart from fixing the broken tests I started with the tests for the newly added channel open notificaion, however it is currently not passing so I commented it out for now while working on somethin else.

* Exposing channel open failure event and adding a channel open failure notification

* Fixes type error caused by the possibility of txConfirmationBlock beeing undefined

* Extending notification card tests with test for block count

* dApp: Add tests for notification channel open

* Changelog entry