dapp: fix serve-pr tooling via Nginx config
Reviewing PRs of the dApp via the serve-pr tool was not working anymore.
This issue was caused by trying to access the favicon.ico resource per
default at the servers root URL. Since there is no such resource, this
lead to errors at the Nginx server. The reason for it is that the server
has/can have multiple versions of the dApp and serves it depending on
the URL query parameter.
As solution the Nginx configuration has been adopted to specifically
handle this case of the web-client, basically ignoring this issue and
let the server continue to serve the dApp.