dapp: fix routing to account if not connected
If the user is not connected with the dApp, he should be still able to
access the account route (and its children). Due to the recent switch to
use a global navigation guard to deny access without being connected,
the account routes have become protected as well.
Therefore the guard for the connection state has been simply extended to
allow navigations to any account route, no matter the connection state.

While extending the tests, it has shown that those are not properly
mocked. So their order was important to make them complete. This has
been solved by tweaking with the store. But it should be target of
future improvements to mock the store completely for unit tests, not
only here. Unfortunately are the getters of the store read-only
protected. So this improvement will need some more attention to find the
best solution.