dapp: fix navigation to transfer steps route
Navigating to the transfer steps route was inheriting the token
parameter of the token route so far. This does not work anymore for all
cases, since the transfer route has the token parameter as optional now.
In such case, the programmatic navigation must provide the exact token
parameter by hand.
This behavior was kinda uncertain before. It was setting the requirement
to always navigate from transfer to transfer steps, or provide a full
path manually.