dapp: add new token list (item) component(s)
As there are multiple places where a list of tokens get shown, it makes
sense to introduce a new component that provides that feature. Therefore
the TokeList and the inner used TokenListItem components have been
created. The list is looking the same as the list for the
SelectTokenRoute so far. But it includes some optimizations for the
template and styling to make it less code and more efficient.