dapp: make use of new token list component
This mainly fixes the issue that if there are too many tokens connected
by the user, the list of the TokenOverlay component was floating the
screen. By using the better written TokenList component, this is not an
issue anymore.
As site effect the styling of the token lists between the
SelectTokenRoute and the TokenOverlay have become equal.
Furthermore has the model to control if the TokenOverlay should be shown
or not become removed from the component itself. Rather the parent
should directly control the rendering state. This also makes the child
component more simple.