dapp: simplify token overlay template and styling
First there was a lot of styling remaining that was meant for the former
token list that has become replaced as its own component. Anyways there
were a couple of more optimizations. Theoretically the banner/button to
add a connect a new token is still pretty ugly and could be improved
even more, but it will probably require a re-design.