dapp: fix incomplete Raiden service mocks in tests
There were a mass of red warning output by the unit tests. To not
desensitize the developer and make actual warning getting missed, this
should be fixed. After all it is just about a two functions that need to
be mocked for the Raiden service within a couple of unit test suites.