dapp: add CI job for e2e tests
The end-to-end tests for the dApp should run in the continuous
integration environment. Therefore a first version of a new job has got
created. This runs the tests with Cypress and stores the report,
screenshots and videos.
This job is so far running within the base executor, chained after
the dApp built. Also the job got integrated into the default workflow to
test it while under development. Both will change in future when using
the integration executor and running nightly.