dapp: add cypress as direct dev dependency
The vue-cli plugin for Cypress is kinda hiding it. This will lead to
issues in the continuous integration environment. Therefore the cypress
package itself got added to the development dependencies of the dApp.
Important to note is that they need to match exactly the version
expected by the vue-cli plugin. This version is pretty ancient already,
but there is an open issue upstream to fix that. We will should update
as soon as this has been fixed.
Additionally the CI will need access to the cached Cypress binary. On
this old version of the package it is required to call the install
command manually. To ensure the installation this got added to the
test:e2e script. If it is already installed this does nothing.