Tx history rounded corners and mobile adjustments
Somhow the bottom rounded corners got lost for the gray box containing the transaction history. In the process of fixing that I also adjusted so that the entire content of the transfer componentes is crollable on mobile. By doing so all the information can be viewed regardless of phone size which was not the case previously.