sdk: switch e2e and integration tests
Historically the behavior of the both test groups have changed and
driven apart. After all the current integration tests are actually
end-to-end tests and vice versa. This also appears in the CI setup where
the default workflow executes the unit and end-to-end tests but the
integration tests are run nightly. The integration directory starts to
differ between the SDK integration tests and the dApp end-to-end tests.
But they both use the same environment. After all this are all
indicators that we should switch the tests. In the end it is just
renaming directories and where they get used.