Improve end-to-end test scripts and docs
Since the SDK as well as the dApp use full setup environment for their
end-to-end tests, there is no more need to differ between them. Thereby
the all scripts around running the tests could have been merged into as
single one. Additionally the quality of this script became slightly
improved. Checkout the script itself to see some details. One major
"breaking" change is that the tests executed depend on the current
working directory where the tests get executed.

Furthermore was the documentation updated. Including the new focus with
not integration related anymore but end-to-end, as well as how to use
the script now. Furthermore the naming was changed consequently. Future
file and directory renaming will follow in upcoming commits to make it
easier to follow the history in the version control system.