Rename integration directory to e2e-environment
The concept/idea/principle behind this directory has changed and is
fully focused on providing and environment for the end-to-end tests of
all Light Client components. This has been already changed in the
content of this directory. Changing the directory name just reflects
those changes. Furthermore it was intended to do this as a separate
commit to have an easier history to see moved files and changed content
as well.