dapp: run e2e test nightly in correct environment
So far the end-to-end tests do not require a full featured environment.
But they will need soon when starting to interact with the blockchain
for example. Therefore the executor for the according CI job was in need
to change. Also do the tests not get executed with the default workflow
anymore, but rather with the nightlies as for the SDK.
Doing so gave the hint that the current job for running the SDK
end-to-end tests is not well defined. Instead of having custom commands
to install dependencies and build the bundle, it should be based on the
already existing jobs doing this. The exchange of workspaces should do
the job to bring the data into the other executor.